Move the bootloader tests to a new file
There are too many tests for the Storage module, so let's move the
bootloader-related tests to a new file.
